feat: implement criterion benchmarks, update python example script
