--- src/lib/mod.rs
@@ -1,7 +1,0 @@
-pub mod client;
-pub mod config;
-pub mod constants;
-pub mod db;
-pub mod errors;
-pub mod handlers;
-pub mod models;
